+++ src/Pages/PageTagDetail.php
@@ -30,6 +30,9 @@ class PageTagDetail extends PageWithDetail
             $this->getFilters($instance);
         } elseif ($this->request->get('tree')) {
             $this->getHierarchy($instance);
+        } elseif ($this->request->get('extra')) {
+            // show extra information without books
+            $this->getExtra($instance);
         } else {
             $this->getEntries($instance);
         }
@@ -53,6 +56,7 @@ class PageTagDetail extends PageWithDetail
         $booklist = new BookList($this->request);
         [$this->entryArray, $this->totalNumber] = $booklist->getBooksByInstanceOrChildren($instance, $this->n);
         $this->sorted = $booklist->orderBy ?? "sort";
+        $this->getExtra($instance);
     }
 
     /**
@@ -65,5 +69,6 @@ class PageTagDetail extends PageWithDetail
         $booklist = new BookList($this->request);
         [$this->entryArray, $this->totalNumber] = $booklist->getBooksByInstance($instance, $this->n);
         $this->sorted = $booklist->orderBy ?? "sort";
+        $this->getExtra($instance);
     }
 }